cli: config and argument refactor

+++ cli/config/robot/robot.go
@@ -1,10 +1,13 @@
 package robot
 
 import (
+	"bytes"
+	"fmt"
 	"os"
 
-	"github.com/robocorp/robo/cli/config/pyproject"
-	yaml "gopkg.in/yaml.v2"
+	"github.com/robocorp/robo/cli/config"
+	"github.com/robocorp/robo/cli/paths"
+	yaml "gopkg.in/yaml.v3"
 )
 
 type RobotYaml struct {
@@ -25,28 +28,49 @@ type Task struct {
 	Command []string `yaml:"command,omitempty"`
 }
 
-func NewFromConfig(cfg pyproject.Robo) *RobotYaml {
-	return &RobotYaml{
+func NewFromConfig(cfg config.Config) RobotYaml {
+	return RobotYaml{
 		Ignored:    []string{".gitignore"},
 		Path:       []string{"."},
 		Pythonpath: []string{"."},
-		Artifacts:  getOutputPath(cfg),
+		Artifacts:  cfg.OutputDir,
 	}
 }
 
-func (it *RobotYaml) SaveAs(path string) error {
-	content, err := yaml.Marshal(it)
+func (it RobotYaml) Encode() ([]byte, error) {
+	var b bytes.Buffer
+	e := yaml.NewEncoder(&b)
+	e.SetIndent(2)
+
+	if err := e.Encode(it); err != nil {
+		return nil, err
+	}
+
+	return b.Bytes(), nil
+}
+
+func (it RobotYaml) Pretty() string {
+	if content, err := it.Encode(); err == nil {
+		return string(content)
+	} else {
+		return fmt.Sprintf("%v", it)
+	}
+}
+
+func (it *RobotYaml) SaveAs(path string, force bool) error {
+	exists, err := paths.Exists(path)
+	if err != nil {
+		return err
+	}
+
+	if exists && !force {
+		return fmt.Errorf("File already exists: %v", path)
+	}
+
+	content, err := it.Encode()
 	if err != nil {
 		return err
 	}
 
 	return os.WriteFile(path, content, 0o666)
-}
-
-func getOutputPath(cfg pyproject.Robo) string {
-	if cfg.Output != "" {
-		return cfg.Output
-	} else {
-		return "output"
-	}
 }
